Added quench (thirst) button
