+++ src/components/Buttons.css
@@ -1,0 +1,27 @@
+.feedbutton {
+    width: 60px;
+    height: 60px;
+    border-radius: 50%;
+    background-position: center;
+    background-image: url('../Assets/feed.jpg');
+    background-size: 60px 60px;
+    margin-right: 75px;
+}
+
+.feedbutton:hover {
+    border: 2px solid red;
+}
+
+.quenchbutton {
+    width: 60px;
+    height: 60px;
+    border-radius: 50%;
+    background-position: center;
+    background-image: url('../Assets/quench.png');
+    background-size: 60px 60px;
+    margin-right: 0px;
+}
+
+.quenchbutton:hover {
+    border: 2px solid red;
+}
