+++ src/components/DogImage.js
@@ -2,11 +2,11 @@ import React from 'react'
 import sad_dog from '../Assets/sad_dog.png'
 import happy_dog from '../Assets/happy_dog.png'
 
-const DogImage = ({hungerValue}) => {
+const DogImage = ({bars}) => {
     return (
         <div style={{textAlign: 'center'}}>
             <div id='container'>
-                { hungerValue < 50 ? <img src={sad_dog} alt='sad_dog.png' class='dog' width="auto" height="200"/> : 
+                { bars.hungerValue < 50 || bars.hungerValue < 50 ? <img src={sad_dog} alt='sad_dog.png' class='dog' width="auto" height="200"/> : 
                 <img src={happy_dog} alt='happy_dog.png' class='dog' width="auto" height="200"/>}
             </div>
             <br />
