--- src/components/feedbutton.css
@@ -1,13 +1,0 @@
-.button {
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    background-position: center;
-    background-image: url('../Assets/feed.jpg');
-    background-size: 60px 60px;
-    margin-right: 75px;
-}
-
-.button:hover {
-    border: 2px solid red;
-}
